refactor: move CmosRtcTimeSourceResolver to emuplat module
